contact_form/requirements.txt
django-widget-tweaks
wagtail-django-recaptcha
django-recaptcha==3.0.0